pyproject.toml
[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[tool.poetry]
name = "dephell_pythons"
version = "0.1.1"
authors = ["orsinium"]
description = "Work with python versions"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.5"
attrs = "*"
dephell_specifier = "*"
packaging = "*"
